+++ server/src/controllers/ColorController/createColorController.js
@@ -1,4 +1,4 @@
-const { Color } = require("../db");
+const { Color } = require("../../db");
 
 const createColor = async (name, description) => {
   const newColor = await Color.create({
