+++ server/src/controllers/ColorController/getColorController.js
@@ -1,4 +1,4 @@
-const { Color } = require("../db.js");
+const { Color } = require("../../db.js");
 
 const findAllTypes = async () => {
   const allColorsArray = await Color.findAll({
