+++ server/src/controllers/LoginController/loginController.js
@@ -1,5 +1,5 @@
 // getUserController.js
-const { User } = require("../db.js");
+const { User } = require("../../db.js");
 const { Op } = require("sequelize");
 const bcrypt = require("bcrypt");
