+++ server/src/controllers/UserControllers/getUserController.js
@@ -1,7 +1,7 @@
 // getUserController.js
-const { User } = require("../db.js");
+const { User } = require("../../db.js");
 const { Op } = require("sequelize");
-const { generateUserToken } = require("../middleware/generateUserToken.js");
+const { generateUserToken } = require("../../middleware/generateUserToken.js");
 
 const findUser = async (userName, email) => {
   console.log("Nombre de usuario:", userName);
