+++ server/src/handlers/ProductHandlers/postProductHandler.js
@@ -1,4 +1,4 @@
-const {postProductController} = require("../../controllers/postProductController");
+const {postProductController} = require("../../controllers/ProductController/postProductController");
 
 const postProductHandler = async (req, res) => {
   const requestData = req.body;
