+++ server/src/handlers/UserHandlers/postUserHandler.js
@@ -1,6 +1,6 @@
 //postUserHandler.js
 
-const { createUser } = require("../../controllers/postUserController");
+const { createUser } = require("../../controllers/UserControllers/postUserController");
 
 const postUserHandler = async (req, res) => {
   try {
